requirements.txt
Django==2.2.0
Pillow>=5.3.0,<5.4.0
django-crispy-forms
djangorestframework
